[별희] 마우스 드래그 앤 드롭으로 3D 모델뷰 위치 변경

+++ Source/MetaAuction/UI/ItemInfo/MAItemDisplayer.h
@@ -64,10 +64,10 @@ private:
 	FVector2D MoveDestination;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Meta = (AllowPrivateAccess = "true"))
-	int32 MoveSpeed;
+	float MoveSpeed;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Meta = (AllowPrivateAccess = "true"))
-	int32 MoveInterpSpeed;
+	float MoveInterpSpeed;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Meta = (AllowPrivateAccess = "true"))
 	FRotator RotationRate;
